requirements.txt
streamlit
torch
torchvision
torchaudio
opencv-python
Pillow
numpy
pandas
yolov5
